requirements.txt
# Core libraries
torch==2.3.0
torchvision==0.18.0
faiss-cpu==1.8.0
numpy>=1.26
pandas>=2.2
tqdm>=4.66
pillow>=10.3


fastapi>=0.111
uvicorn[standard]>=0.29
pillow>=10.3      # already present, but keep for clarity


# Optional (use only if needed)
# streamlit>=1.35
# fastapi>=0.111
# uvicorn[standard]>=0.29
